requirements.txt
pandas
scikit-learn
xgboost
networkx
matplotlib
joblib
python-dotenv